Common report processor for NorSpis

Makes reports for NorSpis typically used for auto reports such as subscriptions, dispatchments and bulletins. As such, please be warned that any changes to this function might render existing auto reports nonfunctional as they are based on static calls based on any previous version of this function. Changes should therefore be thoroughly tested against existing auto reports. Altering the names of the arguments will likely be a breaking change. Adding new arguments should be safe as long as they are provided a default value.

reportProcessor(
  report,
  outputType = "pdf",
  title = "",
  author = "unknown author",
  orgName = "unknown organization",
  orgId = 999999,
  registryName = "ablanor",
  userFullName = "unknown person name",
  userRole = "unknown role",
  userOperator = "unknown operator"
)

Arguments

report
Character string identifying the report to be processed by this function.

outputType
Character string with output format. Must be one of c("html", "pdf"). Defaults to "pdf".

title
Character string giving the report title. Empty string by default.

author
Character string providing the report author. Default value is "unknown author".

orgName
Character string with the name of the organization/hospital. Default is "unknown organization".

orgId
Integer (?) with the id of the organization/hospital. Default is 999999.

registryName
Character string with registry name. Default is "norspis".

userFullName
Character string giving the person name, normally the user requesting the report. Default is "unknown person name".

userRole
Character string giving a user role, normally the one of the user requesting the report. Default is "unknown role".

userOperator
Character string with some name of an operator, whatever that is... Default is "unknown operator".

Value

A character string with a path to where the produced file is located.

Examples

## Make the start page for NorSpis
reportFilePath <- reportProcessor(report = "veiledning",
                                  outputType = "html",
                                  title = "Example report")
#> 
#> 
#> processing file: veiledning.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#> 
#> output file: veiledning.knit.md
#> /usr/local/bin/pandoc +RTS -K512m -RTS veiledning.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cRPjx/11411d00945a.html --lua-filter /Users/runner/work/_temp/Library/bookdown/rmarkdown/lua/custom-environment.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --metadata-file /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cRPjx/file11418729e73 --self-contained --wrap preserve --variable bs3=TRUE --standalone --section-divs --template /Users/runner/work/_temp/Library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --include-in-header /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cRPjx/rmarkdown-str11411634a998.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 
#> 
#> Output created: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp2cRPjx/11411d00945a.html